refactor: remove deprecated handlers, tests, and markdown processing code; introduce middleware for logging and security
