+++ internal/server/middleware.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"fmt"
@@ -6,14 +6,14 @@ import (
 )
 
 // middleware outputs the method and uri the request is hitting on the server
-func (app *application) logRequests(next http.Handler) http.Handler {
+func (s *Server) logRequests(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		fmt.Printf("-> %s | %s \n", r.Method, r.URL)
 		next.ServeHTTP(w, r)
 	})
 }
 
-func (app *application) setCommon(next http.Handler) http.Handler {
+func (s *Server) setCommon(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Content-Type", "text/html")
 		next.ServeHTTP(w, r)
@@ -34,7 +34,7 @@ func (app *application) setCommon(next http.Handler) http.Handler {
 // middleware so these headers are present on responses from downstream handlers.
 // This does not provide a Content-Security-Policy or HSTS; add those headers
 // separately where required.
-func (app *application) setSecurity(next http.Handler) http.Handler {
+func (s *Server) setSecurity(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("X-Frame-Options", "deny")
 		w.Header().Set("X-Content-Type-Options", "nosniff")
@@ -45,9 +45,9 @@ func (app *application) setSecurity(next http.Handler) http.Handler {
 	})
 }
 
-func (app *application) logHandler(handlerName string, next http.Handler) http.Handler {
+func (s *Server) logHandler(handlerName string, next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		app.logger.Info("handling request",
+		s.logger.Info("handling request",
 			"method", r.Method,
 			"url", r.URL.Path,
 			"handler", handlerName)
